% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c50.R
\name{c5_one_language}
\alias{c5_one_language}
\title{Compute C5.0 Probabilities}
\usage{
c5_one_language(df, score, ediff, c5_object = read_c50_eng())
}
\arguments{
\item{df}{a data.frame}

\item{score}{column with fuzzy score between two food items}

\item{ediff}{column with relative percentage of absolute energy_difference}

\item{c5_object}{(default is the English Eurofir mapping classifier}
}
\value{
data.frame
}
\description{
Compute C5.0 probabilities using the C5.O tree from food score and energy differences
}
\author{
Marc Lamarine
}
