% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c50.R
\name{c5_predict}
\alias{c5_predict}
\title{Predict the mapping plausability with 0 or 1 using a C5.0 classification tree}
\usage{
c5_predict(df, fuzzy_score_col, energy_diff_col, c5_object = read_c50_eng(),
  type = "prob")
}
\arguments{
\item{df}{data.frame}

\item{fuzzy_score_col}{column with the string distance score}

\item{energy_diff_col}{column with the percentage of energy difference}

\item{c5_object}{C5.0 tree classifier}

\item{type}{type of prediction output}
}
\value{
a vector
}
\description{
Predict the mapping plausability with 0 or 1 using a C5.0 classification tree
}
\author{
Marc Lamarine
}
\keyword{internal}
