% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{food_score}
\alias{food_score}
\title{Food Comparisons Table with fuzzy scores}
\format{data.frame object}
\usage{
data(food_score)
}
\description{
Food Comparisons Table with fuzzy scores
}
\keyword{datasets}
