% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchers.R
\name{read_c50_eng}
\alias{read_c50_eng}
\title{Fetch the English Food Name Classifier}
\usage{
read_c50_eng()
}
\value{
a C5.0 object
}
\description{
Fetch the English Food Name Classifier
}
\author{
Marc Lamarine
}
