% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchers.R
\name{read_c50_ori}
\alias{read_c50_ori}
\title{Fetch the Original Food Name Classifier}
\usage{
read_c50_ori()
}
\value{
a C5.0 object
}
\description{
Fetch the Original Food Name Classifier
}
\author{
Marc Lamarine
}
